[ROFO-182] RankingQueryService, RankingCommandService 생성, 메소드 이동
