+++ src/main/kotlin/kr/weit/roadyfoody/ranking/presentation/api/RankingController.kt
@@ -2,7 +2,7 @@ package kr.weit.roadyfoody.ranking.presentation.api
 
 import jakarta.validation.constraints.Positive
 import kr.weit.roadyfoody.foodSpots.application.dto.UserReportCount
-import kr.weit.roadyfoody.foodSpots.application.service.FoodSpotsCommandService
+import kr.weit.roadyfoody.ranking.application.service.RankingQueryService
 import kr.weit.roadyfoody.ranking.presentation.spec.RankingControllerSpec
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
@@ -12,12 +12,12 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 @RequestMapping("/api/v1/ranking")
 class RankingController(
-    private val foodSpotsCommandService: FoodSpotsCommandService,
+    private val rankingQueryService: RankingQueryService,
 ) : RankingControllerSpec {
     @GetMapping("/report")
     override fun getReportRanking(
         @Positive(message = "size는 양수여야 합니다.")
         @RequestParam(defaultValue = "10")
         size: Long,
-    ): List<UserReportCount> = foodSpotsCommandService.getReportRanking(size)
+    ): List<UserReportCount> = rankingQueryService.getReportRanking(size)
 }
