+++ src/main/kotlin/kr/weit/roadyfoody/ranking/utils/RankingConstans.kt
@@ -1,0 +1,4 @@
+package kr.weit.roadyfoody.ranking.utils
+
+const val REPORT_RANKING_KEY = "rofo:user-report-ranking"
+const val REPORT_RANKING_UPDATE_LOCK = "updateReportRankingLock"
